+++ examples/quadratic_implicit.py
@@ -16,9 +16,33 @@
 #
 #
 # This work has been cleared for public release, distribution unlimited, case
-# number: AFRL-2023-XXXX. The views expressed are those of the author and do not
-# necessarily reflect the official policy or position of the Department of the
-# Air Force, the Department of Defense, or the U.S. government.
+# number: AFRL-2023-5713.
+#
+# The views expressed are those of the authors and do not reflect the
+# official guidance or position of the United States Government, the
+# Department of Defense or of the United States Air Force.
+#
+# Statement from DoD: The Appearance of external hyperlinks does not
+# constitute endorsement by the United States Department of Defense (DoD) of
+# the linked websites, of the information, products, or services contained
+# therein. The DoD does not exercise any editorial, security, or other
+# control over the information you may find at these locations.
+from concurrent import futures
+import grpc
+import philote_mdo.general as pmdo
+from philote_mdo.examples import QuadradicImplicit
 
 
-pmdo.run_server(RemoteQuadratic())
+def run():
+    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
+    discipline = pmdo.ImplicitServer(discipline=QuadradicImplicit())
+    discipline.attach_to_server(server)
+
+    server.add_insecure_port("[::]:50051")
+    server.start()
+    print("Server started. Listening on port 50051.")
+    server.wait_for_termination()
+
+
+if __name__ == "__main__":
+    run()