+++ philote_mdo/examples/paraboloid.py
@@ -16,7 +16,7 @@
 #
 #
 # This work has been cleared for public release, distribution unlimited, case
-# number: AFRL-2023-XXXX.
+# number: AFRL-2023-5713.
 #
 # The views expressed are those of the authors and do not reflect the
 # official guidance or position of the United States Government, the
@@ -29,7 +29,6 @@
 # control over the information you may find at these locations.
 import philote_mdo.general as pmdo
 
-class Parabaloid(pmdo.ExplicitServer):
 
 class Paraboloid(pmdo.ExplicitDiscipline):
     """