+++ philote_mdo/examples/quadratic.py
@@ -16,7 +16,7 @@
 #
 #
 # This work has been cleared for public release, distribution unlimited, case
-# number: AFRL-2023-XXXX.
+# number: AFRL-2023-5713.
 #
 # The views expressed are those of the authors and do not reflect the
 # official guidance or position of the United States Government, the
@@ -27,19 +27,23 @@
 # the linked websites, of the information, products, or services contained
 # therein. The DoD does not exercise any editorial, security, or other
 # control over the information you may find at these locations.
+import numpy as np
 import philote_mdo.general as pmdo
 
 
 class QuadradicImplicit(pmdo.ImplicitDiscipline):
     def setup(self):
-        self.define_input("a", shape=(1,))
-        self.define_input("b", shape=(1,))
-        self.define_input("c", shape=(1,))
+        self.add_input("a", shape=(1,))
+        self.add_input("b", shape=(1,))
+        self.add_input("c", shape=(1,))
 
-        self.define_output("x", shape=(1,))
+        self.add_output("x", shape=(1,))
 
     def setup_partials(self):
-        self.define_partials("x", "*")
+        self.declare_partials("x", "a")
+        self.declare_partials("x", "b")
+        self.declare_partials("x", "c")
+        self.declare_partials("x", "x")
 
     def compute_residuals(self, inputs, outputs, residuals):
         a = inputs["a"]
@@ -53,18 +57,18 @@ class QuadradicImplicit(pmdo.ImplicitDiscipline):
         a = inputs["a"]
         b = inputs["b"]
         c = inputs["c"]
-        outputs["x"] = (-b + (b**2 - 4 * a * c) ** 0.5) / (2 * a)
+        outputs["x"] = np.array([(-b + (b**2 - 4 * a * c) ** 0.5) / (2 * a)])
 
-    def linearize(self, inputs, outputs, partials):
+    def residual_partials(self, inputs, outputs, partials):
         a = inputs["a"]
         b = inputs["b"]
         c = inputs["c"]
         x = outputs["x"]
 
-        partials["x", "a"] = x**2
-        partials["x", "b"] = x
-        partials["x", "c"] = 1.0
-        partials["x", "x"] = 2 * a * x + b
+        partials["x", "a"] = np.array([x**2])
+        partials["x", "b"] = np.array([x])
+        partials["x", "c"] = np.array([1.0])
+        partials["x", "x"] = np.array([2 * a * x + b])
 
         self.inv_jac = 1.0 / (2 * a * x + b)