+++ philote_mdo/general/implicit_server.py
@@ -16,7 +16,7 @@
 #
 #
 # This work has been cleared for public release, distribution unlimited, case
-# number: AFRL-2023-XXXX.
+# number: AFRL-2023-5713.
 #
 # The views expressed are those of the authors and do not reflect the
 # official guidance or position of the United States Government, the
@@ -33,16 +33,22 @@ import philote_mdo.general as pmdo
 from philote_mdo.utils import get_chunk_indices
 
 
-class ImplicitServer(pmdo.DisciplineServer, disc.ImplicitDisciplineServicer):
+class ImplicitServer(pmdo.DisciplineServer, disc.ImplicitServiceServicer):
     """
     Base class for creating an implicit discipline server.
     """
 
-    def __init__(self):
-        super().__init__()
-        self._implicit = True
+    def __init__(self, discipline=None):
+        super().__init__(discipline=discipline)
 
-    def Residuals(self, request_iterator, context):
+    def attach_to_server(self, server):
+        """
+        Attaches this discipline server class to a gRPC server.
+        """
+        super().attach_to_server(server)
+        disc.add_ImplicitServiceServicer_to_server(self, server)
+
+    def ComputeResiduals(self, request_iterator, context):
         """
         Computes the residuals and sends the results to the client.
         """
@@ -53,24 +59,23 @@ class ImplicitServer(pmdo.DisciplineServer, disc.ImplicitDisciplineServicer):
         flat_outputs = {}
         residuals = {}
 
-        # preallocate the inputs for the implicit discipline
         self.preallocate_inputs(inputs, flat_inputs, outputs, flat_outputs)
         self.process_inputs(request_iterator, flat_inputs, flat_outputs)
 
-        # call the user-defined compute function
+        # call the user-defined compute_residuals function
         self._discipline.compute_residuals(inputs, outputs, residuals)
 
         for res_name, value in residuals.items():
-            for b, e in get_chunk_indices(value.size, self.num_double):
+            for b, e in get_chunk_indices(value.size, self._stream_opts.num_double):
                 yield data.Array(
                     name=res_name,
                     start=b,
                     end=e,
                     type=data.kResidual,
-                    continuous=value.ravel()[b:e],
+                    data=value.ravel()[b:e],
                 )
 
-    def Solve(self, request_iterator, context):
+    def SolveResiduals(self, request_iterator, context):
         """
         Solves the implicit discipline so that the residuals are driven to zero.
         """
@@ -81,19 +86,22 @@ class ImplicitServer(pmdo.DisciplineServer, disc.ImplicitDisciplineServicer):
         flat_outputs = {}
 
         self.preallocate_inputs(inputs, flat_inputs, outputs, flat_outputs)
-
         self.process_inputs(request_iterator, flat_inputs, flat_outputs)
 
-        # call the user-defined compute function
+        # call the user-defined solve function
         self._discipline.solve_residuals(inputs, outputs)
 
         for output_name, value in outputs.items():
-            for b, e in get_chunk_indices(value.size, self.num_double):
+            for b, e in get_chunk_indices(value.size, self._stream_opts.num_double):
                 yield data.Array(
-                    name=output_name, start=b, end=e, continuous=value.ravel()[b:e]
+                    name=output_name,
+                    start=b,
+                    end=e,
+                    type=data.kOutput,
+                    data=value.ravel()[b:e],
                 )
 
-    def ResidualGradients(self, request_iterator, context):
+    def ComputeResidualGradients(self, request_iterator, context):
         """
         Computes the residual gradients and sends the results to the client.
         """
@@ -103,22 +111,22 @@ class ImplicitServer(pmdo.DisciplineServer, disc.ImplicitDisciplineServicer):
         outputs = {}
         flat_outputs = {}
 
-        # preallocate the input and discrete input arrays
         self.preallocate_inputs(inputs, flat_inputs, outputs, flat_outputs)
         jac = self.preallocate_partials()
         self.process_inputs(request_iterator, flat_inputs, flat_outputs)
 
-        # call the user-defined compute_partials function
-        self._discipline.linearize(inputs, jac)
+        # call the user-defined residual partials function
+        self._discipline.residual_partials(inputs, outputs, jac)
 
         for jac, value in jac.items():
-            for b, e in get_chunk_indices(value.size, self.num_double):
+            for b, e in get_chunk_indices(value.size, self._stream_opts.num_double):
                 yield data.Array(
                     name=jac[0],
                     subname=jac[1],
+                    type=data.kPartial,
                     start=b,
                     end=e,
-                    continuous=value.ravel()[b:e],
+                    data=value.ravel()[b:e],
                 )
 
     # def MatrixFreeGradients(self, request_iterator, context):