+++ test/run_tests.py
@@ -16,7 +16,7 @@
 #
 #
 # This work has been cleared for public release, distribution unlimited, case
-# number: AFRL-2023-XXXX.
+# number: AFRL-2023-5713.
 #
 # The views expressed are those of the authors and do not reflect the
 # official guidance or position of the United States Government, the
@@ -33,9 +33,11 @@ from test_discipline import TestDiscipline
 
 from test_discipline_server import TestDisciplineServer
 from test_explicit_server import TestExplicitServer
+from test_implicit_server import TestImplicitServer
 
 from test_discipline_client import TestDisciplineClient
 from test_explicit_client import TestExplicitClient
+from test_implicit_client import TestImplicitClient
 
 
 if __name__ == "__main__":